Move color selector window inside the prefs GUI
Move GlassServerControlGui_ColorSelector to become a child of GlassServerControlGui_Window, leaving the contents of the GUI otherwise unchanged.

This fixes issues with users being unable to hit "escape" to close the prefs menu, and it also locks the color selector window to the prefs window, preventing it from becoming hidden behind the main window.